--- src/hr/job/model/entity/Job.java
@@ -1,20 +1,0 @@
-package hr.job.model.entity;
-
-import common.BaseEntity;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import java.time.LocalDateTime;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class Job extends BaseEntity {
-    private int jobId;
-    private String jobName;
-    private String jobCode;
-    private String description;
-}
