--- src/hr/userAccount/controller/UserAccountController.java
@@ -1,4 +1,0 @@
-package hr.userAccount.controller;
-
-public class UserAccountController {
-}
