--- src/hr/userAccount/model/entity/UserAccount.java
@@ -1,21 +1,0 @@
-package hr.userAccount.model.entity;
-
-import common.BaseEntity;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import java.time.LocalDateTime;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class UserAccount extends BaseEntity {
-    private int accountId;
-    private String userId;
-    private String password;
-
-    private int employeeId;
-}
