+++ src/humanResource/job/model/entity/Job.java
@@ -1,0 +1,18 @@
+package humanResource.job.model.entity;
+
+import common.BaseEntity;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class Job extends BaseEntity {
+    private int jobId;
+    private String jobName;
+    private String jobCode;
+    private String description;
+}
