+++ src/humanResource/userAccount/controller/UserAccountController.java
@@ -1,0 +1,4 @@
+package humanResource.userAccount.controller;
+
+public class UserAccountController {
+}
